+++ config/saito_config.php
@@ -113,6 +113,14 @@ $config['Saito']['Settings']['uploader'] = (new UploaderConfig())
      */
     ->setDefaultMaxFileSize('8MB')
     /**
+     * Threshold file size (and rough target) for resizing images (jpeg/png)
+     */
+    ->setDefaultMaxResize('650kB')
+    /**
+     * Image quality factor when resizing images (integer between 0 and 100)
+     */
+    ->setImageCompressionQuality(92)
+    /**
      * Allowed mime/types
      */
     ->addType('audio/mpeg')